exracting viewmodeldecoration code into a separate file
