+++ src/vs/editor/browser/view/renderingContext.ts
@@ -6,7 +6,8 @@
 import { Position } from '../../common/core/position.js';
 import { Range } from '../../common/core/range.js';
 import { ViewportData } from '../../common/viewLayout/viewLinesViewportData.js';
-import { IViewLayout, ViewModelDecoration } from '../../common/viewModel.js';
+import { IViewLayout } from '../../common/viewModel.js';
+import { ViewModelDecoration } from '../../common/viewModel/viewModelDecoration.js';
 
 export interface IViewLines {
 	linesVisibleRangesForRange(range: Range, includeNewLines: boolean): LineVisibleRanges[] | null;
