+++ src/vs/editor/common/viewLayout/viewLinesViewportData.ts
@@ -5,7 +5,8 @@
 
 import { Range } from '../core/range.js';
 import { Selection } from '../core/selection.js';
-import { IPartialViewLinesViewportData, IViewModel, IViewWhitespaceViewportData, ViewLineRenderingData, ViewModelDecoration } from '../viewModel.js';
+import { IPartialViewLinesViewportData, IViewModel, IViewWhitespaceViewportData, ViewLineRenderingData } from '../viewModel.js';
+import { ViewModelDecoration } from '../viewModel/viewModelDecoration.js';
 
 /**
  * Contains all data needed to render at a specific viewport.
