+++ src/vs/editor/common/viewModel/viewModelDecoration.ts
@@ -1,0 +1,79 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { IModelDecoration, IModelDecorationOptions, ITextModel } from '../model.js';
+import { Range } from '../core/range.js';
+import { StandardTokenType } from '../encodedTokenAttributes.js';
+
+export class ViewModelDecoration {
+	_viewModelDecorationBrand: void = undefined;
+
+	public readonly range: Range;
+	public readonly options: IModelDecorationOptions;
+
+	constructor(range: Range, options: IModelDecorationOptions) {
+		this.range = range;
+		this.options = options;
+	}
+}
+
+export function isModelDecorationVisible(model: ITextModel, decoration: IModelDecoration): boolean {
+	if (decoration.options.hideInCommentTokens && isModelDecorationInComment(model, decoration)) {
+		return false;
+	}
+
+	if (decoration.options.hideInStringTokens && isModelDecorationInString(model, decoration)) {
+		return false;
+	}
+
+	return true;
+}
+
+export function isModelDecorationInComment(model: ITextModel, decoration: IModelDecoration): boolean {
+	return testTokensInRange(
+		model,
+		decoration.range,
+		(tokenType) => tokenType === StandardTokenType.Comment
+	);
+}
+
+export function isModelDecorationInString(model: ITextModel, decoration: IModelDecoration): boolean {
+	return testTokensInRange(
+		model,
+		decoration.range,
+		(tokenType) => tokenType === StandardTokenType.String
+	);
+}
+
+/**
+ * Calls the callback for every token that intersects the range.
+ * If the callback returns `false`, iteration stops and `false` is returned.
+ * Otherwise, `true` is returned.
+ */
+function testTokensInRange(model: ITextModel, range: Range, callback: (tokenType: StandardTokenType) => boolean): boolean {
+	for (let lineNumber = range.startLineNumber; lineNumber <= range.endLineNumber; lineNumber++) {
+		const lineTokens = model.tokenization.getLineTokens(lineNumber);
+		const isFirstLine = lineNumber === range.startLineNumber;
+		const isEndLine = lineNumber === range.endLineNumber;
+
+		let tokenIdx = isFirstLine ? lineTokens.findTokenIndexAtOffset(range.startColumn - 1) : 0;
+		while (tokenIdx < lineTokens.getCount()) {
+			if (isEndLine) {
+				const startOffset = lineTokens.getStartOffset(tokenIdx);
+				if (startOffset > range.endColumn - 1) {
+					break;
+				}
+			}
+
+			const callbackResult = callback(lineTokens.getStandardTokenType(tokenIdx));
+			if (!callbackResult) {
+				return false;
+			}
+			tokenIdx++;
+		}
+	}
+	return true;
+}
+
